requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.27.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
flask>=3.0.0
anthropic>=0.25.0
twilio>=9.0.0
